refactor(UI-Home): 拆分首页组件为独立模块，清理未使用的类型定义;
- 将 HomeBody 从 page.tsx 迁移至独立的 home-body.tsx，page.tsx 缩减为薄入口
- AGENT_ID / APP_NAME 改为具名导出，支持测试直接引用
- 删除 route.ts 中未使用的 AguiLifecycleEvent 局部类型定义
- 集成测试 import 路径同步更新
